Adding Game interface and GameState sealed class
We can simplify our set logic by having a game interface that allows us to interact with a standard and tie break game in the same fashion. I added the GameState sealed class to represent the state of the game. This can be useful internally for the Game implementation and for the client (Set) implementation. A sealed class is a good choice here as it acts like an enum allowing us to declare a number of fixed states the game can be in while also associating data with each state, such as the score once in progress and the winner of the game once complete.