Adding Standard Game implementation
Initial implementation simply keeps count of all player points and calculates state and score based on number of points scored for each player.

Whenever a player scores a point, we firstly check if the game is not complete, if so we ignore the call - we could potentially throw an illegal state exception here. Next we update our running count of player points won based on the provided player, before calculating the new GameState.

The first thing we do do calculate the sate is check if anyone has won, if not we need to determine if the game has reached deuce. If not we can simply use the player points won tallies as index to get the scores from our PRE_DEUCE_SCORES array. if we have reached deuce, the player with the most points will have Advantage, or if points are equal we are at deuce