Adding Tie Break implementation
Works in a similar way to a standard game, but there is no need to translate the points scored into 15,30,40,deuce etc. The logic for the game is almost identical except the min win points to win is the different and the logic for printing the scores is simplified. We could probably refactor the two games to share some similar logic, either through composition or inheritance