Adding Set Specificiation
This was a little trickier to to write tests for without a mocking framework as I had no way to stub the Game implementations and my tests had to rely on the concrete implementations of Standard and Tiebreak games. This had probably resulted in some double testing of those classes.