Adding Set Implementation
Shared a lot of code with the StandardGame and TieBreak classes here but there are more subtle differences. I thought about storing the game instances that each player had won as this could be a good source of information, however at this point it was not needed to so I decided to use an Int to determine the number of games each player had won.

After that it was simply a case of updating the way we managed the state and changing the logic for whether a set had been won, followed by which game should be pulled next.

I considered providing factories to create new instances of the Game interfaces to help with testing but without a mocking framework that could verify calls to the mocks I decided it was not worth it (as i would have had to roll my own)