Unifying Game / Tiebreak logic
The further you look into it the more you realise that a standard and tiebreak game are actually the same game, the only difference is the number of points required to win and the way the scores are read. Because of this we no longer need to make the winner method abstract and can simply ask subclasses to provide both players and the minimum number of points to win, leaving each class to simply provide some configuration and describe the score.

The benefit of having this logic in the abstract class is that it makes it obvious that the two games are essentially the same, it also helps keep our code DRY. The downside is flexibility in other implementations, however I think its fair to assume that the rules of tennis are unlikely to change very often.